refactor: move achievement components out of route module
